NFC: create a separate Python module to prepare future changes
