+++ build_tools/benchmarks/run_benchmarks_on_android.py
@@ -14,26 +14,6 @@ It expects that `adb` is installed, and there is iree tools cross-compiled
 towards Android. If to capture traces, another set of tracing-enabled iree
 tools and the Tracy `capture` tool should be cross-compiled towards Android.
 
-It also expects the benchmark artifacts are generated by building the
-`iree-benchmark-suites` target in the following directory structure:
-
-<root-build-dir>/benchmark_suites
-└── <benchmark-category> (e.g., TFLite)
-    ├── <benchmark-suite> (e.g., MobileBertSquad-fp32)
-    │   ├── <benchmark-case> (e.g., iree-vulkan__GPU-Mali-Valhall__kernel-execution)
-    │   │   ├── tool
-    │   │   └── flagfile
-    │   ├── ...
-    │   │   ├── tool
-    │   │   └── flagfile
-    │   └── <benchmark_case>
-    │       ├── tool
-    │       └── flagfile
-    └── vmfb
-        ├── <compiled-iree-model>-<sha1>.vmfb
-        ├── ...
-        └── <compiled-iree-model>-<sha1>.vmfb
-
 Example usages:
 
   # Without trace generation
@@ -66,9 +46,10 @@ from common.benchmark_definition import (
     AndroidDeviceInfo, BenchmarkInfo, BenchmarkResults, BenchmarkRun,
     execute_cmd, execute_cmd_and_get_output, get_android_device_model,
     get_android_gpu_name, IREE_PRETTY_NAMES_TO_DRIVERS)
+from common.benchmark_suite import (BENCHMARK_SUITE_REL_PATH,
+                                    compose_info_object,
+                                    filter_benchmarks_for_category)
 
-# All benchmarks' relative path against root build directory.
-BENCHMARK_SUITE_REL_PATH = "benchmark_suites"
 # VMFB files' relative path against a benchmark category directory.
 VMFB_REL_PATH = "vmfb"
 
@@ -218,106 +199,6 @@ def adb_start_cmd(cmd_args: Sequence[str],
   return subprocess.Popen(cmd, stdout=subprocess.PIPE, universal_newlines=True)
 
 
-def compose_benchmark_info_object(device_info: AndroidDeviceInfo,
-                                  benchmark_category_dir: str,
-                                  benchmark_case_dir: str) -> BenchmarkInfo:
-  """Creates an BenchmarkInfo object to describe the benchmark.
-
-  Args:
-    device_info: an AndroidDeviceInfo object.
-    benchmark_category_dir: the directory to a specific benchmark category.
-    benchmark_case_dir: a directory containing the benchmark case.
-
-  Returns:
-    A BenchmarkInfo object.
-  """
-  # Extract the model name from the directory path. This uses the relative
-  # path under the root model directory. If there are multiple segments,
-  # additional ones will be placed in parentheses.
-  model_name = os.path.relpath(benchmark_case_dir, benchmark_category_dir)
-  # Now we have <model-name>/.../<iree-driver>__<target-arch>__<bench_mode>,
-  # Remove the last segment.
-  model_name = os.path.dirname(model_name)
-  main, rest = os.path.split(model_name)
-  if main:
-    # Tags coming from directory structure.
-    model_name = main
-    model_tags = [re.sub(r"\W+", "-", rest)]
-  else:
-    # Tags coming from the name itself.
-    model_name, rest = rest.split("-", 1)
-    model_tags = rest.split(",")
-
-  # Extract benchmark info from the directory path following convention:
-  #   <iree-driver>__<target-architecture>__<benchmark_mode>
-  root_immediate_dir = os.path.basename(benchmark_case_dir)
-  iree_driver, target_arch, bench_mode = root_immediate_dir.split("__")
-
-  model_source = os.path.basename(benchmark_category_dir)
-
-  return BenchmarkInfo(model_name=model_name,
-                       model_tags=model_tags,
-                       model_source=model_source,
-                       bench_mode=bench_mode.split(","),
-                       runner=iree_driver,
-                       device_info=device_info)
-
-
-def filter_benchmarks_for_category(benchmark_category_dir: str,
-                                   cpu_target_arch: str,
-                                   gpu_target_arch: str,
-                                   driver_filter: Optional[str],
-                                   verbose: bool = False) -> Sequence[str]:
-  """Filters benchmarks in a specific category for the given device.
-
-  Args:
-    benchmark_category_dir: the directory to a specific benchmark category.
-    cpu_target_arch: CPU target architecture.
-    gpu_target_arch: GPU target architecture.
-    driver_filter: only run benchmarks for the given driver if not None.
-    verbose: whether to print additional debug info.
-
-  Returns:
-    A list containing all matched benchmark cases' directories.
-  """
-  matched_benchmarks = []
-
-  # Go over all benchmarks in the model directory to find those matching the
-  # current Android device's CPU/GPU architecture.
-  for root, dirs, _ in os.walk(benchmark_category_dir):
-    # Take the immediate directory name and try to see if it contains compiled
-    # models and flagfiles. This relies on the following directory naming
-    # convention:
-    #   <iree-driver>__<target-architecture>__<benchmark_mode>
-    root_immediate_dir = os.path.basename(root)
-    segments = root_immediate_dir.split("__")
-    if len(segments) != 3 or not segments[0].startswith("iree-"):
-      continue
-
-    iree_driver, target_arch, bench_mode = segments
-    iree_driver = iree_driver[len("iree-"):].lower()
-    target_arch = target_arch.lower()
-
-    # We can choose this benchmark if it matches the driver and CPU/GPU
-    # architecture.
-    matched_driver = (driver_filter is None or
-                      iree_driver == driver_filter.lower())
-    matched_arch = (target_arch == cpu_target_arch or
-                    target_arch == gpu_target_arch)
-    should_choose = matched_driver and matched_arch
-    if should_choose:
-      matched_benchmarks.append(root)
-
-    if verbose:
-      print(f"dir: {root}")
-      print(f"  iree_driver: {iree_driver}")
-      print(f"  target_arch: {target_arch}")
-      print(f"  bench_mode: {bench_mode}")
-      print(f"  chosen: {should_choose}")
-
-  return matched_benchmarks
-
-
 def run_benchmarks_for_category(
     device_info: AndroidDeviceInfo,
     benchmark_category_dir: str,
@@ -393,9 +274,8 @@ def run_benchmarks_for_category(
     with open(os.path.join(benchmark_case_dir, MODEL_TOOLFILE_NAME)) as f:
       tool = f.read().strip()
 
-    benchmark_info = compose_benchmark_info_object(device_info,
-                                                   benchmark_category_dir,
-                                                   benchmark_case_dir)
+    benchmark_info = compose_info_object(device_info, benchmark_category_dir,
+                                         benchmark_case_dir)
     benchmark_key = str(benchmark_info)
     # If we're not running the benchmark or the capture, just skip ahead.
     # No need to push files.
@@ -652,7 +532,7 @@ def parse_arguments():
   parser.add_argument("--output",
                       "-o",
                       default=None,
-                      help="Path to the ouput file")
+                      help="Path to the output file")
   parser.add_argument("--capture_tarball",
                       "--capture-tarball",
                       default=None,
